devices: fs: Cosmetic change for the order of cast and bitwise operation
Follow-up of the comment of https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5740485/comment/4043e33d_785f0b63/.
Casting before bitwise operation looks a bit safer.

No functional change.

BUG=none
TEST=CQ

Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5772144
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>

+++ devices/src/virtio/fs/passthrough.rs
@@ -2382,7 +2382,7 @@ impl FileSystem for PassthroughFs {
                 data,
                 name,
                 entry.inode,
-                (flags & !(libc::O_CREAT | libc::O_EXCL | libc::O_NOCTTY)) as u32,
+                flags as u32 & !((libc::O_CREAT | libc::O_EXCL | libc::O_NOCTTY) as u32),
             )
             .map_err(|e| {
                 // Don't leak the entry.
